Extracted code associated with user settings file into separate object and converted credentials handling to use Maven's effective settings resolution in `maven-settings` library.
